Created the Main Menu
Created the main menu function as well as a general purpose menu function.
