+++ menus.h
@@ -1,0 +1,20 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/cFiles/file.h to edit this template
+ */
+
+/**@file menus.h
+ * @author Diogo Coelho
+ * @version 1.0
+ * @date 6 de dezembro de 2023, 14:18
+ */
+
+#ifndef MENUS_H
+#define MENUS_H
+
+int menu(int num_op, char *title, char *op_array[]);
+
+void main_menu();
+
+#endif /* MENUS_H */
+
